Refactor massive tests blob in persistent_engine into sub-modules
