+++ .jules/atlas.md
@@ -15,3 +15,6 @@
 ## 2024-05-19 - Sub-modularizing the Tests Blob
 **Tangle:** The `relvar-storage/src/storage/heap/tests.rs` file had grown back into a 3,000+ line Blob despite being extracted from the main source file, making it extremely difficult to navigate and maintain.
 **Blueprint:** Removed the `tests.rs` monolith and replaced it with a `tests/` directory structure containing specialized sub-modules (e.g., `insert.rs`, `scan.rs`, `update.rs`, `delete.rs`, `gc.rs`, `corruption.rs`, and `version.rs`) along with a `common.rs` file for shared helpers. This restores high cohesion to the testing structure.
+## 2024-05-19 - Sub-modularizing the Persistent Engine Tests Blob
+**Tangle:** The `relvar-storage/src/persistent_engine/tests.rs` file had grown into a massive Blob containing nearly 1,400 lines of inline tests for various aspects of the persistent engine (transactions, recovery, garbage collection, and basic operations).
+**Blueprint:** Removed the `tests.rs` monolith and extracted its contents into a specialized sub-module structure under a new `relvar-storage/src/persistent_engine/tests/` directory. Created explicit sub-modules for `basic.rs`, `transaction.rs`, `recovery.rs`, and `gc.rs`, along with a `common.rs` for shared test initialization.
