+++ relvar-storage/src/persistent_engine/tests/common.rs
@@ -1,0 +1,10 @@
+use crate::persistent_engine::*;
+use relvar_core::types::{ScalarType, TupleType};
+
+pub fn test_rel_type() -> RelationType {
+    RelationType::new(
+        TupleType::new()
+            .with_attribute("id", ScalarType::Int)
+            .with_attribute("name", ScalarType::String),
+    )
+}
