+++ relvar-storage/src/persistent_engine/tests/mod.rs
@@ -1,0 +1,7 @@
+#![cfg(test)]
+
+mod basic;
+mod common;
+mod gc;
+mod recovery;
+mod transaction;
